Change port to class property of UdpThread
Change port from a public member to a class property with only an
accessor, preventing the user from changing the port after
instantiation. UdpThread is used a member for the UdpSensor and
UdpActuator classes for abstracting UDP communication and it is not
expected to change between transmitting and receiving or even ports in
these classes.

Add port property to UdpSensor.