Simplify input torque conversion
Instead of reading a character, adding it to a string, converting the
string to a chracter array, and then performing the float conversion,
use a character array to store input.

Remove multiple value logic as only a single value is ever sent. To
transmit more values, sending bits instead of the character
representation would be more efficient and should replace the current
method.